Mihael's components added
